Use pure-scala code instead of Enum
