+++ airframe-msgpack/shared/src/main/scala/wvlet/airframe/msgpack/io/BufferUnpacker.scala
@@ -32,11 +32,11 @@ class BufferUnpacker {
   def lastReadByteLength: Int = _lastReadByteLength
 
   private def unexpected(expectedCode: String, actual: Byte) = {
-    val f = Code.messageFormatOf(actual)
+    val f = MessageFormat.of(actual)
     if (f == MessageFormat.NEVER_USED) {
       throw new MessagePackException(NEVER_USED_FORMAT, s"Expected ${expectedCode}, but found 0xC1 (NEVER_USED) byte")
     } else {
-      val name     = f.getValueType.name()
+      val name     = f.valueType.name
       val typeName = name.substring(0, 1) + name.substring(1).toLowerCase(Locale.ENGLISH)
       throw new MessagePackException(INVALID_TYPE, f"Expected ${expectedCode}, but got ${typeName} (${actual}%02x)")
     }
