+++ airframe-msgpack/shared/src/main/scala/wvlet/airframe/msgpack/spi/Code.scala
@@ -75,108 +75,4 @@ object Code {
   val MAP32: Byte            = 0xdf.toByte
   val NEGFIXINT_PREFIX: Byte = 0xe0.toByte
 
-  private[spi] val formatTable = new Array[MessageFormat](256)
-
-  {
-    // Preparing a look up table for converting byte values into MessageFormat types
-    for (b <- 0 to 0xFF) {
-      val mf = toMessageFormat(b.toByte)
-      formatTable(b) = mf
-    }
-  }
-
-  /**
-    * Returns a MessageFormat type of the specified byte value
-    *
-    * @param code MessageFormat of the given byte
-    * @return
-    */
-  def messageFormatOf(code: Byte): MessageFormat = formatTable(code & 0xFF)
-
-  /**
-    * Converting a byte value into MessageFormat. For faster performance, use {@link #valueOf}
-    *
-    * @param b MessageFormat of the given byte
-    * @return
-    */
-  private[spi] def toMessageFormat(b: Byte): MessageFormat = {
-    if (Code.isPosFixInt(b))
-      MessageFormat.POSFIXINT
-    else if (Code.isNegFixInt(b))
-      MessageFormat.NEGFIXINT
-    else if (Code.isFixStr(b))
-      MessageFormat.FIXSTR
-    else if (Code.isFixedArray(b))
-      MessageFormat.FIXARRAY
-    else if (Code.isFixedMap(b))
-      MessageFormat.FIXMAP
-    else {
-      b match {
-        case Code.NIL =>
-          MessageFormat.NIL
-        case Code.FALSE =>
-          MessageFormat.BOOLEAN
-        case Code.TRUE =>
-          MessageFormat.BOOLEAN
-        case Code.BIN8 =>
-          MessageFormat.BIN8
-        case Code.BIN16 =>
-          MessageFormat.BIN16
-        case Code.BIN32 =>
-          MessageFormat.BIN32
-        case Code.EXT8 =>
-          MessageFormat.EXT8
-        case Code.EXT16 =>
-          MessageFormat.EXT16
-        case Code.EXT32 =>
-          MessageFormat.EXT32
-        case Code.FLOAT32 =>
-          MessageFormat.FLOAT32
-        case Code.FLOAT64 =>
-          MessageFormat.FLOAT64
-        case Code.UINT8 =>
-          MessageFormat.UINT8
-        case Code.UINT16 =>
-          MessageFormat.UINT16
-        case Code.UINT32 =>
-          MessageFormat.UINT32
-        case Code.UINT64 =>
-          MessageFormat.UINT64
-        case Code.INT8 =>
-          MessageFormat.INT8
-        case Code.INT16 =>
-          MessageFormat.INT16
-        case Code.INT32 =>
-          MessageFormat.INT32
-        case Code.INT64 =>
-          MessageFormat.INT64
-        case Code.FIXEXT1 =>
-          MessageFormat.FIXEXT1
-        case Code.FIXEXT2 =>
-          MessageFormat.FIXEXT2
-        case Code.FIXEXT4 =>
-          MessageFormat.FIXEXT4
-        case Code.FIXEXT8 =>
-          MessageFormat.FIXEXT8
-        case Code.FIXEXT16 =>
-          MessageFormat.FIXEXT16
-        case Code.STR8 =>
-          MessageFormat.STR8
-        case Code.STR16 =>
-          MessageFormat.STR16
-        case Code.STR32 =>
-          MessageFormat.STR32
-        case Code.ARRAY16 =>
-          MessageFormat.ARRAY16
-        case Code.ARRAY32 =>
-          MessageFormat.ARRAY32
-        case Code.MAP16 =>
-          MessageFormat.MAP16
-        case Code.MAP32 =>
-          MessageFormat.MAP32
-        case _ =>
-          MessageFormat.NEVER_USED
-      }
-    }
-  }
 }
