--- airframe-msgpack/shared/src/main/scala/wvlet/airframe/msgpack/spi/MessageFormat.java
@@ -1,74 +1,0 @@
-/*
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package wvlet.airframe.msgpack.spi;
-
-/**
- * Describes the list of the message format types defined in the MessagePack specification.
- * TODO: Convert this into Scala
- */
-public enum MessageFormat
-{
-    // INT7
-    POSFIXINT(ValueType.INTEGER),
-    // MAP4
-    FIXMAP(ValueType.MAP),
-    // ARRAY4
-    FIXARRAY(ValueType.ARRAY),
-    // STR5
-    FIXSTR(ValueType.STRING),
-    NIL(ValueType.NIL),
-    NEVER_USED(null),
-    BOOLEAN(ValueType.BOOLEAN),
-    BIN8(ValueType.BINARY),
-    BIN16(ValueType.BINARY),
-    BIN32(ValueType.BINARY),
-    EXT8(ValueType.EXTENSION),
-    EXT16(ValueType.EXTENSION),
-    EXT32(ValueType.EXTENSION),
-    FLOAT32(ValueType.FLOAT),
-    FLOAT64(ValueType.FLOAT),
-    UINT8(ValueType.INTEGER),
-    UINT16(ValueType.INTEGER),
-    UINT32(ValueType.INTEGER),
-    UINT64(ValueType.INTEGER),
-
-    INT8(ValueType.INTEGER),
-    INT16(ValueType.INTEGER),
-    INT32(ValueType.INTEGER),
-    INT64(ValueType.INTEGER),
-    FIXEXT1(ValueType.EXTENSION),
-    FIXEXT2(ValueType.EXTENSION),
-    FIXEXT4(ValueType.EXTENSION),
-    FIXEXT8(ValueType.EXTENSION),
-    FIXEXT16(ValueType.EXTENSION),
-    STR8(ValueType.STRING),
-    STR16(ValueType.STRING),
-    STR32(ValueType.STRING),
-    ARRAY16(ValueType.ARRAY),
-    ARRAY32(ValueType.ARRAY),
-    MAP16(ValueType.MAP),
-    MAP32(ValueType.MAP),
-    NEGFIXINT(ValueType.INTEGER);
-
-    private final ValueType valueType;
-
-    private MessageFormat(ValueType valueType)
-    {
-        this.valueType = valueType;
-    }
-
-    public ValueType getValueType() { return valueType; }
-
-}
